rename "common" module to "chill" + fix circular deps
