+++ src/app/app.component.html
@@ -1,6 +1,6 @@
 <ion-app>
   <ion-split-pane contentId="main-content">
-    <ion-menu>
+    <ion-menu contentId="main-content">
       <ion-header>
         <ion-toolbar>
           <ion-title>Menu</ion-title>
