+++ src/app/app.component.ts
@@ -3,8 +3,8 @@ import { Component } from '@angular/core';
 import { Platform } from '@ionic/angular';
 import { SplashScreen } from '@ionic-native/splash-screen/ngx';
 import { StatusBar } from '@ionic-native/status-bar/ngx';
-import { AppTitleService } from './common/services/app-title.service';
-import { AppPaths } from './app-routing.module';
+import { AppTitleService } from './chill/services/app-title.service';
+import { AppPaths } from './app.constants';
 
 @Component({
   selector: 'app-root',
