+++ src/app/app.constants.ts
@@ -1,0 +1,5 @@
+export const AppPaths = {
+    SHOPPER: 'shopper',
+    TEMPLATES: 'templates'
+  };
+
