+++ src/app/chill/title-bar/title-bar.component.ts
