+++ src/app/shopper/shopper-routing.module.ts
@@ -3,10 +3,7 @@ import { RouterModule, Routes } from '@angular/router';
 import { ShopperPage } from './shopper.page';
 import { ShoplistManagerComponent } from './components/shoplist-manager/shoplist-manager.component';
 import { ShoplistEditionComponent } from './components/shoplist-edition/shoplist-edition.component';
-
-export const ShopperPaths = {
-    EDIT: 'edit'
-};
+import { ShopperPaths } from './shopper.constants';
 
 const routes: Routes = [
     {
