+++ src/app/shopper/shopper.constants.ts
@@ -1,0 +1,3 @@
+export const ShopperPaths = {
+    EDIT: 'edit'
+};
