+++ src/app/templates/templates.module.ts
@@ -4,7 +4,7 @@ import { CommonModule } from '@angular/common';
 import { TemplatesRoutingModule } from './templates-routing.module';
 import { TemplatesPageComponent } from './templates-page/templates-page.component';
 import { IonicModule } from '@ionic/angular';
-import { ChillModule } from '../common/chill.module';
+import { ChillModule } from '../chill/chill.module';
 
 
 @NgModule({
